--- pointfog/.flowconfig
@@ -1,2 +1,0 @@
-[ignore]
-.*/node_modules/config-chain/test/broken.json
