+++ pointfog/run-tests.sh
@@ -1,2 +1,1 @@
 eslint -c .eslintrc.js ./
-flow
