--- pointfog/src/_build_info.js
@@ -1,5 +1,0 @@
-export default {
-        "commit": "100019291ecffec3114a145b1db3837ef9198bc7",
-        "user": "mateljk1",
-        "creation": new Date(1599666003819),
-    };
