--- pointfog/src/_config.js
@@ -1,11 +1,0 @@
-import _build_info from "./_build_info";
-
-window._build_info = _build_info;
-
-const Config = {
-
-    // Default URL for public-facing colocard server
-    colocardUrl: "https://colocard.thebossdev.io",
-};
-
-export default Config;
