+++ pointfog/src/colors.ts
@@ -1,7 +1,5 @@
-// @flow
+
 
 const LINE_COLOR: [number, number, number] = [0, 230, 250];
 
-export {
-    LINE_COLOR,
-};
+export { LINE_COLOR };
